requirements.txt
flask
flask-cors
gunicorn
python-dotenv
requests